refactor(lib/oauth): move crypto to lib; brick re-exports
